[all]: Refactor configs to MOM_common
Because we can read a sequence of input files using the input.nml
directive 'parameter_filename', we refactor out all the configuration
options that are common across all experiments into a single MOM_common.
Configuration that is specific to each experiment is then in MOM_input,
and the variants of each experiment go in MOM_override as usual.