[eddies]: RHO input generation/configuration
The mk_input.py script was re-written to output the same configuration
for a (continuous) isopycnal experiment. This involves a constant
temperature configuration, but varying layer thicknesses in order to
obtain the same z-space initial condition as previously.

We also add z-space diagnostic output using remapped diagnostics to make
sure the outputs are interpretable (RPE calculation should be able to
use the native output).