[waves]: More frequent output, and interface heights
Because a lot of this test case revolves around how layers evolve with
time (e.g. with different regridding time scales), we need the layer
data in our diagnostic output. Particularly, we need the interface
positions so that we can plot the results in depth-space.

Timestep was also bumped up because it's easily stable at 1800s.