[analysis]: Allow RPE calculation to use topography
Previously, the RPE calculation code assumes a flat-bottomed domain,
without taking into account the effect of topography. It was assumed
(based on some minimal evidence) that accounting for topography wouldn't
significantly affect the calculated RPE, for the cost in the algorithm
change. To allow for a more rigorous verification of this, it's possible
to specify a 'topod' parameter for the 'calc_rpe' function.