[writing]: Use LaTeX packages and make single document
We add a preamble.md which contains options for the document, such as
title, authors, and LaTeX packages to use. We use cleveref, caption and
siunitx packages to make writing significantly easier. When building the
document, specify all the markdown files to be concatenated
together (there's probably a way to automate this with
spacemacs/whatever).